# specs/001-phase2-web/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon II Todo API
  description: API for the Hackathon II Todo full-stack web application
  version: 1.0.0
  contact:
    name: API Support
    email: support@todoapp.com

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server

paths:
  /api/tasks:
    get:
      summary: List user's tasks
      description: Retrieve a list of tasks for the authenticated user with optional filtering and sorting
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter tasks by completion status
          required: false
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
        - name: sort
          in: query
          description: Sort tasks by creation date or title
          required: false
          schema:
            type: string
            enum: [created, title]
            default: created
        - name: order
          in: query
          description: Sort order (ascending or descending)
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of user's tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '500':
          description: Internal server error
    post:
      summary: Create a new task
      description: Create a new task for the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: Task title
                  example: "Buy groceries"
                description:
                  type: string
                  description: Task description
                  example: "Milk, eggs, bread"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '500':
          description: Internal server error

  /api/tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by ID for the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '404':
          description: Task not found
        '500':
          description: Internal server error
    put:
      summary: Update a task
      description: Update an existing task for the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Task title
                  example: "Buy groceries"
                description:
                  type: string
                  description: Task description
                  example: "Milk, eggs, bread"
                completed:
                  type: boolean
                  description: Task completion status
                  example: false
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '404':
          description: Task not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a task
      description: Delete a specific task for the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '404':
          description: Task not found
        '500':
          description: Internal server error

  /api/tasks/{id}/complete:
    patch:
      summary: Toggle task completion
      description: Toggle the completion status of a specific task
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Task completion status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '404':
          description: Task not found
        '500':
          description: Internal server error

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        user_id:
          type: string
          description: ID of the user who owns the task
          example: "user_12345"
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          description: Task description
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Task completion status
          example: false
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
          example: "2026-01-02T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Task last update timestamp
          example: "2026-01-02T10:00:00Z"

security:
  - BearerAuth: []